+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2024-09-01
+### Feat
+- Added possibility to use a more stable time configuration for the game, that doesn't reset it's configuration after play mode is exited
+- Made it possible to select whether to use a provided "Game" time configuration or use a dev config that doesn't change
+- Added possibility to accelerate / slow time during runtime with a custom inspector on the TimeManager
+- Added a "Save" method to the TimeManager to register the current runtime configuration progress in the used TimeConfigurationSO (only when not in testing mode to avoid overriding the testing config scriptable objects). This method also returns the updated TimeConfigurationSO so it can be plugged into any Save / Load system.
+- Added elements to demo scene for clearer understanding of how to use the external TimeConfiguration vs the Testing one in the TimeManager.
+### Fix
+- Fixed an issue in which events with 0 checks were not triggered
+
 ## [0.6.0] - 2024-09-01
 ### Feat
 - Extended checkers associations in order to select if it should run an AND or OR logic on the checkers associated to the event